Create actual Layout component with title and container
The previous layout component was just a padding and head elements.
I tried making the layout an actual layout now, including title
and content sections. It will also help (maybe) the nav buttons
in the near future.
