+++ client/pages/index.tsx
@@ -1,13 +1,13 @@
-import Layout from "components/layout";
+import Base from "components/base";
 
 const Index = () => (
-  <Layout title="Olá!">
+  <Base title="Olá!">
     <h1 className="uk-heading-medium uk-heading-line uk-text-center">
       <span>
         Porti<span className="uk-text-primary">folio</span>
       </span>
     </h1>
-  </Layout>
+  </Base>
 );
 
 export default Index;
